requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
bcrypt>=4.0.1
python-dotenv>=1.0.0
pydantic>=1.8.2
requests>=2.31.0
python-jwt>=4.0.0